backend/requirements.txt
# --- Core Framework ---
fastapi==0.115.2
uvicorn[standard]==0.32.0

# --- Database & ORM ---
SQLAlchemy==2.0.36
psycopg[binary]==3.2.1
mysqlclient==2.2.5
alembic==1.14.0

# --- Data Validation & Settings ---
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1

# --- Authentication & Security ---
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# --- Utilities & Middleware ---
email-validator==2.2.0
requests==2.32.3
python-multipart==0.0.9

# --- Logging & Monitoring ---
loguru==0.7.2

# --- Testing ---
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.28.1

# --- Optional (for advanced usage) ---
redis==5.1.1              # For caching / sessions
celery==5.4.0             # For async tasks (email, reports)
flower==2.0.1             # Celery monitoring UI
python-dateutil==2.9.0.post0
